make prod deploy work again (move main app logic into __init__.py and add dummy main)
