+++ backend/main.py
@@ -1,0 +1,1 @@
+from xivdpscalc import app
